+++ app/Core/Events/Contracts/EventInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Sendity\Core\Events\Contracts;
+
+interface EventInterface
+{
+}