+++ app/Core/Events/Contracts/ListenerInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Sendity\Core\Events\Contracts;
+
+interface ListenerInterface
+{
+    public function handle(EventInterface $event): void;
+}