+++ bootstrap/app.php
@@ -7,6 +7,7 @@ use Sendity\Controllers\HomeController;
 use Sendity\Services\Logger;
 use Sendity\Core\Exceptions\ExceptionHandler;
 use Sendity\Http\Response;
+use Sendity\Core\Events\EventDispatcher;
 
 require_once __DIR__ . '/../vendor/autoload.php';
 
@@ -38,11 +39,16 @@ $container->singleton(
     ExceptionHandler::class,
     fn () => new ExceptionHandler()
 );
+$container->singleton(
+    EventDispatcher::class,
+    fn ($container) => new EventDispatcher($container)
+);
 // Router singleton
 $container->singleton(
     \Sendity\Http\Router::class,
     fn ($c) => new \Sendity\Http\Router($c)
 );
+
 
 // Shared router instance
 $router = $container->get(\Sendity\Http\Router::class);
@@ -62,7 +68,26 @@ $router->get('/api/status', function () {
 $router->get('/user/{id}', function ($id) {
     return "User ID: {$id}";
 });
+$router->get('/event-test', function () use ($container) {
 
+    $events = $container->get(
+        \Sendity\Core\Events\EventDispatcher::class
+    );
+
+    $events->listen(
+        \Sendity\Events\MailSent::class,
+        \Sendity\Listeners\LogMailSent::class
+    );
+
+    $events->dispatch(
+        new \Sendity\Events\MailSent(
+            'cedric@example.com',
+            'Hello Sendity!'
+        )
+    );
+
+    return 'Event dispatched!';
+});
 
 // Run application
 return $container->get(Application::class);